refactor(updateForProps): rename updateForKeys
